requirements.txt
Flask>=3.0.0,<4
Flask-SQLAlchemy>=3.1.0,<4
python-dateutil>=2.8.2,<3
